requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow
numpy
tensorflow==2.16.1
keras==3.3.3
python-multipart
pydantic==2.7.1
python-dotenv
sqlalchemy==2.0.30
sqlmodel==0.0.16